config.py
import os
from dotenv import load_dotenv

load_dotenv()

class Config:
    SECRET_KEY = os.getenv("FLASK_SECRET_KEY")

    SUPABASE_URL = os.getenv("SUPABASE_URL")
    SUPABASE_KEY = os.getenv("SUPABASE_KEY")

    GEMINI_API_KEY = os.getenv("GEMINI_API_KEY")
    SUPABASE_SERVICE_KEY= os.getenv("SUPABASE_SERVICE_KEY")
