requirements.txt
flask
python-dotenv
supabase
pdfplumber
python-docx
google-genai
supabase
reportlab
